examples/mock-example/py-upload/requirements.txt
cassandra-driver==3.12.0
numpy==1.13.3
pandas==0.21.1
pkg-resources==0.0.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
tqdm==4.19.5
